refactor: extract interface MailService
